Repeatable yantra symlink

+++ yantra/yantra_symlinks.sh
@@ -1,4 +1,17 @@
-sudo ln -s $HOME/Developer/scaffold/yantra/nixos/ /etc/nixos
+#!/usr/bin/env bash
 
-ln -s $HOME/Developer/scaffold/yantra/xmonad $HOME/.xmonad
-ln -s $HOME/Developer/scaffold/yantra/lilyterm $HOME/.config/lilyterm
+if [ ! -L /etc/nixos ]; then
+  rm -rf /etc/nixos 2> /dev/null
+  sudo ln -s $HOME/Developer/scaffold/yantra/nixos/ /etc/nixos
+fi
+
+if [ ! -L $HOME/.xmonad ]; then
+  rm -rf $HOME/.xmonad 2> /dev/null
+  ln -s $HOME/Developer/scaffold/yantra/xmonad $HOME/.xmonad
+fi
+
+if [ ! -L $HOME/.config/lilyterm ]; then
+  rm -rf $HOME/.config/lilyterm/ 2> /dev/null
+  mkdir -p $HOME/.config/
+  ln -s $HOME/Developer/scaffold/yantra/lilyterm $HOME/.config/lilyterm
+fi
